requirements.txt
appdirs==1.4.4
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
cssselect==1.1.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
mysql-connector-python==8.0.30
parse==1.19.0
protobuf==3.20.1
pycodestyle==2.9.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.11
w3lib==1.22.0
websockets==10.3
zipp==3.8.1
